+++ database/migrations/2016_05_10_160918_create_table_alat_bantu_tangkap.php
@@ -21,7 +21,7 @@ class CreateTableAlatBantuTangkap extends Migration
             $table->integer('jumlah')->nullable();
             $table->string('satuan_jumlah', 30)->nullable();
             $table->integer('harga_beli')->nullable();
-            $table->float('umur_teknis')->nullable();
+            $table->double('umur_teknis')->nullable();
             
             $table->softDeletes();
             $table->timestamps();
