+++ database/migrations/2016_05_11_145344_create_table_ops_nelayan.php
@@ -17,8 +17,8 @@ class CreateTableOpsNelayan extends Migration
             $table->integer('id_responden');
             $table->integer('id_master_biaya');
 
-            $table->float('jumlah')->nullable();
-            $table->float('harga_satuan')->nullable();
+            $table->double('jumlah')->nullable();
+            $table->double('harga_satuan')->nullable();
             $table->integer('total_biaya')->nullable();
 
             $table->softDeletes();
